--- a/RecoTracker/LSTCore/src/alpaka/Segment.h
+++ b/RecoTracker/LSTCore/src/alpaka/Segment.h
@@ -192,6 +192,15 @@ namespace ALPAKA_ACCELERATOR_NAMESPACE::lst {
                                                          float dPhiChange,
                                                          float dPhiChangeMin,
                                                          float dPhiChangeMax,
+#ifdef CUT_VALUE_DEBUG
+                                                         float zHi,
+                                                         float zLo,
+                                                         float rtHi,
+                                                         float rtLo,
+                                                         float dAlphaInner,
+                                                         float dAlphaOuter,
+                                                         float dAlphaInnerOuter,
+#endif
                                                          unsigned int idx) {
     segments.mdIndices()[idx][0] = lowerMDIndex;
     segments.mdIndices()[idx][1] = upperMDIndex;
@@ -206,6 +215,16 @@ namespace ALPAKA_ACCELERATOR_NAMESPACE::lst {
     segments.dPhiChanges()[idx] = __F2H(dPhiChange);
     segments.dPhiChangeMins()[idx] = __F2H(dPhiChangeMin);
     segments.dPhiChangeMaxs()[idx] = __F2H(dPhiChangeMax);
+
+#ifdef CUT_VALUE_DEBUG
+    segments.zHis()[idx] = __F2H(zHi);
+    segments.zLos()[idx] = __F2H(zLo);
+    segments.rtHis()[idx] = __F2H(rtHi);
+    segments.rtLos()[idx] = __F2H(rtLo);
+    segments.dAlphaInners()[idx] = __F2H(dAlphaInner);
+    segments.dAlphaOuters()[idx] = __F2H(dAlphaOuter);
+    segments.dAlphaInnerOuters()[idx] = __F2H(dAlphaInnerOuter);
+#endif
   }
 
   template <typename TAcc>
@@ -356,7 +375,18 @@ namespace ALPAKA_ACCELERATOR_NAMESPACE::lst {
                                                                   float& dPhiChange,
                                                                   float& dPhiChangeMin,
                                                                   float& dPhiChangeMax,
+#ifdef CUT_VALUE_DEBUG
+                                                                  float& dAlphaInnerMDSegment,
+                                                                  float& dAlphaOuterMDSegment,
+                                                                  float& dAlphaInnerMDOuterMD,
+                                                                  float& zLo,
+                                                                  float& zHi,
+#endif
                                                                   const float ptCut) {
+#ifndef CUT_VALUE_DEBUG
+    float dAlphaInnerMDSegment, dAlphaOuterMDSegment, dAlphaInnerMDOuterMD;
+    float zLo, zHi;
+#endif
     float xIn, yIn, zIn, rtIn, xOut, yOut, zOut, rtOut;
 
     xIn = mds.anchorX()[innerMDIndex];
@@ -374,9 +404,9 @@ namespace ALPAKA_ACCELERATOR_NAMESPACE::lst {
 
     const float zGeom = modules.layers()[innerLowerModuleIndex] <= 2 ? 2.f * kPixelPSZpitch : 2.f * kStrip2SZpitch;
 
-    float zLo = zIn + (zIn - kDeltaZLum) * (rtOut / rtIn - 1.f) * (zIn > 0.f ? 1.f : dzDrtScale) -
-                zGeom;  //slope-correction only on outer end
-    float zHi = zIn + (zIn + kDeltaZLum) * (rtOut / rtIn - 1.f) * (zIn < 0.f ? 1.f : dzDrtScale) + zGeom;
+    //slope-correction only on outer end
+    zLo = zIn + (zIn - kDeltaZLum) * (rtOut / rtIn - 1.f) * (zIn > 0.f ? 1.f : dzDrtScale) - zGeom;
+    zHi = zIn + (zIn + kDeltaZLum) * (rtOut / rtIn - 1.f) * (zIn < 0.f ? 1.f : dzDrtScale) + zGeom;
 
     if ((zOut < zLo) || (zOut > zHi))
       return false;
@@ -420,9 +450,9 @@ namespace ALPAKA_ACCELERATOR_NAMESPACE::lst {
 
     float innerMDAlpha = mds.dphichanges()[innerMDIndex];
     float outerMDAlpha = mds.dphichanges()[outerMDIndex];
-    float dAlphaInnerMDSegment = innerMDAlpha - dPhiChange;
-    float dAlphaOuterMDSegment = outerMDAlpha - dPhiChange;
-    float dAlphaInnerMDOuterMD = innerMDAlpha - outerMDAlpha;
+    dAlphaInnerMDSegment = innerMDAlpha - dPhiChange;
+    dAlphaOuterMDSegment = outerMDAlpha - dPhiChange;
+    dAlphaInnerMDOuterMD = innerMDAlpha - outerMDAlpha;
 
     float dAlphaInnerMDSegmentThreshold = dAlphaThresholdValues[0];
     float dAlphaOuterMDSegmentThreshold = dAlphaThresholdValues[1];
@@ -449,7 +479,18 @@ namespace ALPAKA_ACCELERATOR_NAMESPACE::lst {
                                                                   float& dPhiChange,
                                                                   float& dPhiChangeMin,
                                                                   float& dPhiChangeMax,
+#ifdef CUT_VALUE_DEBUG
+                                                                  float& dAlphaInnerMDSegment,
+                                                                  float& dAlphaOuterMDSegment,
+                                                                  float& dAlphaInnerMDOuterMD,
+                                                                  float& rtLo,
+                                                                  float& rtHi,
+#endif
                                                                   const float ptCut) {
+#ifndef CUT_VALUE_DEBUG
+    float dAlphaInnerMDSegment, dAlphaOuterMDSegment, dAlphaInnerMDOuterMD;
+    float rtLo, rtHi;
+#endif
     float xIn, yIn, zIn, rtIn, xOut, yOut, zOut, rtOut;
 
     xIn = mds.anchorX()[innerMDIndex];
@@ -481,10 +522,10 @@ namespace ALPAKA_ACCELERATOR_NAMESPACE::lst {
     float dLum = alpaka::math::copysign(acc, kDeltaZLum, zIn);
     float drtDzScale = sdSlope / alpaka::math::tan(acc, sdSlope);
 
-    float rtLo = alpaka::math::max(
-        acc, rtIn * (1.f + dz / (zIn + dLum) * drtDzScale) - rtGeom, rtIn - 0.5f * rtGeom);  //rt should increase
-    float rtHi = rtIn * (zOut - dLum) / (zIn - dLum) +
-                 rtGeom;  //dLum for luminous; rGeom for measurement size; no tanTheta_loc(pt) correction
+    //rt should increase
+    rtLo = alpaka::math::max(acc, rtIn * (1.f + dz / (zIn + dLum) * drtDzScale) - rtGeom, rtIn - 0.5f * rtGeom);
+    //dLum for luminous; rGeom for measurement size; no tanTheta_loc(pt) correction
+    rtHi = rtIn * (zOut - dLum) / (zIn - dLum) + rtGeom;
 
     // Completeness
     if ((rtOut < rtLo) || (rtOut > rtHi))
@@ -544,9 +585,9 @@ namespace ALPAKA_ACCELERATOR_NAMESPACE::lst {
 
     float innerMDAlpha = mds.dphichanges()[innerMDIndex];
     float outerMDAlpha = mds.dphichanges()[outerMDIndex];
-    float dAlphaInnerMDSegment = innerMDAlpha - dPhiChange;
-    float dAlphaOuterMDSegment = outerMDAlpha - dPhiChange;
-    float dAlphaInnerMDOuterMD = innerMDAlpha - outerMDAlpha;
+    dAlphaInnerMDSegment = innerMDAlpha - dPhiChange;
+    dAlphaOuterMDSegment = outerMDAlpha - dPhiChange;
+    dAlphaInnerMDOuterMD = innerMDAlpha - outerMDAlpha;
 
     float dAlphaInnerMDSegmentThreshold = dAlphaThresholdValues[0];
     float dAlphaOuterMDSegmentThreshold = dAlphaThresholdValues[1];
@@ -573,8 +614,21 @@ namespace ALPAKA_ACCELERATOR_NAMESPACE::lst {
                                                             float& dPhiChange,
                                                             float& dPhiChangeMin,
                                                             float& dPhiChangeMax,
+#ifdef CUT_VALUE_DEBUG
+                                                            float& dAlphaInnerMDSegment,
+                                                            float& dAlphaOuterMDSegment,
+                                                            float& dAlphaInnerMDOuterMD,
+                                                            float& zLo,
+                                                            float& zHi,
+                                                            float& rtLo,
+                                                            float& rtHi,
+#endif
                                                             const float ptCut) {
     if (modules.subdets()[innerLowerModuleIndex] == Barrel and modules.subdets()[outerLowerModuleIndex] == Barrel) {
+#ifdef CUT_VALUE_DEBUG
+      rtLo = -999.f;
+      rtHi = -999.f;
+#endif
       return runSegmentDefaultAlgoBarrel(acc,
                                          modules,
                                          mds,
@@ -588,8 +642,19 @@ namespace ALPAKA_ACCELERATOR_NAMESPACE::lst {
                                          dPhiChange,
                                          dPhiChangeMin,
                                          dPhiChangeMax,
+#ifdef CUT_VALUE_DEBUG
+                                         dAlphaInnerMDSegment,
+                                         dAlphaOuterMDSegment,
+                                         dAlphaInnerMDOuterMD,
+                                         zLo,
+                                         zHi,
+#endif
                                          ptCut);
     } else {
+#ifdef CUT_VALUE_DEBUG
+      zLo = -999.f;
+      zHi = -999.f;
+#endif
       return runSegmentDefaultAlgoEndcap(acc,
                                          modules,
                                          mds,
@@ -603,6 +668,13 @@ namespace ALPAKA_ACCELERATOR_NAMESPACE::lst {
                                          dPhiChange,
                                          dPhiChangeMin,
                                          dPhiChangeMax,
+#ifdef CUT_VALUE_DEBUG
+                                         dAlphaInnerMDSegment,
+                                         dAlphaOuterMDSegment,
+                                         dAlphaInnerMDOuterMD,
+                                         rtLo,
+                                         rtHi,
+#endif
                                          ptCut);
     }
   }
@@ -643,6 +715,10 @@ namespace ALPAKA_ACCELERATOR_NAMESPACE::lst {
 
             float dPhi, dPhiMin, dPhiMax, dPhiChange, dPhiChangeMin, dPhiChangeMax;
 
+#ifdef CUT_VALUE_DEBUG
+            float zLo, zHi, rtLo, rtHi, dAlphaInnerMDSegment, dAlphaOuterMDSegment, dAlphaInnerMDOuterMD;
+#endif
+
             unsigned int innerMiniDoubletAnchorHitIndex = mds.anchorHitIndices()[innerMDIndex];
             unsigned int outerMiniDoubletAnchorHitIndex = mds.anchorHitIndices()[outerMDIndex];
             dPhiMin = 0;
@@ -662,6 +738,15 @@ namespace ALPAKA_ACCELERATOR_NAMESPACE::lst {
                                       dPhiChange,
                                       dPhiChangeMin,
                                       dPhiChangeMax,
+#ifdef CUT_VALUE_DEBUG
+                                      dAlphaInnerMDSegment,
+                                      dAlphaOuterMDSegment,
+                                      dAlphaInnerMDOuterMD,
+                                      zLo,
+                                      zHi,
+                                      rtLo,
+                                      rtHi,
+#endif
                                       ptCut)) {
               unsigned int totOccupancySegments =
                   alpaka::atomicAdd(acc,
@@ -692,6 +777,15 @@ namespace ALPAKA_ACCELERATOR_NAMESPACE::lst {
                                    dPhiChange,
                                    dPhiChangeMin,
                                    dPhiChangeMax,
+#ifdef CUT_VALUE_DEBUG
+                                   zHi,
+                                   zLo,
+                                   rtHi,
+                                   rtLo,
+                                   dAlphaInnerMDSegment,
+                                   dAlphaOuterMDSegment,
+                                   dAlphaInnerMDOuterMD,
+#endif
                                    segmentIdx);
               }
             }