--- a/RecoTracker/LSTCore/standalone/code/core/AnalysisConfig.h
+++ b/RecoTracker/LSTCore/standalone/code/core/AnalysisConfig.h
@@ -114,9 +114,6 @@ public:
   // Boolean to write lower level objects
   bool do_lower_level;
 
-  // Boolean to write gnn ntuple
-  bool gnn_ntuple;
-
   // String to hold the MAKETARGET setting from compile
   std::string compilation_target;
 
@@ -132,8 +129,44 @@ public:
   // Boolean to disable pLS duplicate cleaning
   bool no_pls_dupclean;
 
+  // Boolean to enable MD branches
+  bool md_branches;
+
+  // Boolean to enable LS branches
+  bool ls_branches;
+
+  // Boolean to enable T3 branches
+  bool t3_branches;
+
+  // Boolean to enable T5 branches
+  bool t5_branches;
+
+  // Boolean to enable pLS branches
+  bool pls_branches;
+
+  // Boolean to enable pT3 branches
+  bool pt3_branches;
+
+  // Boolean to enable pT5 branches
+  bool pt5_branches;
+
+  // Boolean to enable occupancy branches
+  bool occ_branches;
+
+  // Boolean to enable T3 DNN branches
+  bool t3dnn_branches;
+
+  // Boolean to enable T5 DNN branches
+  bool t5dnn_branches;
+
+  // Boolean to enable pT3 DNN branches
+  bool pt3dnn_branches;
+
   // Boolean to enable jet branches
   bool jet_branches;
+
+  // Boolean to enable extra sim branches
+  bool extra_sim_branches;
 };
 
 extern AnalysisConfig ana;