@ariostas 2025-02-24
@Hoobidoobidoo instead of adding additional options and flags, I decided to use the existing `CUT_VALUE_DEBUG` flag. So to enable the extra outputs you can just use the `-d` flag. Let me know if this would be an issue for you.

I verified that it compiles (after fixing a couple of typos I made), but could you also verify that the output ntuple contains all the data you want and that it looks good?