@slava77 2025-02-24
was there a feature change in this PR in the algorithm ( I thought this was just ntuple-related)
why is the efficiency changing?
https://raw.githubusercontent.com/SegmentLinking/TrackLooper-plots-archive/cmssw/PR149_43311a5_standalone/eff_pt_comp.png