@ariostas 2025-02-24
> was there a feature change in this PR in the algorithm ( I thought this was just ntuple-related)
> why is the efficiency changing?

No, there was no change in the algorithm. There must be a bug somewhere