@ariostas 2025-04-01
The cpu timings on cgpu-1 also look good.

```
This PR (357dc9f)
Total Timing Summary
Average time for map loading = 349.674 ms
Average time for input loading = 8143.17 ms
Average time for lst::Event creation = 0.00399635 ms
   Evt    Hits       MD       LS      T3       T5       pLS       pT5      pT3      TC       Reset    Event     Short             Rate
   avg     19.8    281.5    141.6    143.3     33.5    363.8     77.8     93.2    107.7      1.1    1263.3     879.7+/- 220.8    1264.3   explicit[s=1]
   avg     19.3    280.3    139.6    146.2     32.8    359.7     76.4     92.6    106.9      0.9    1254.8     875.7+/- 212.7     329.5   explicit[s=4]
   avg     20.8    291.4    145.0    147.8     34.9    370.2     80.0     95.7    109.6      1.1    1296.3     905.3+/- 217.0      94.9   explicit[s=16]
   avg     22.6    295.4    146.6    149.3     35.8    374.6     80.6     97.7    111.2      1.0    1314.9     917.7+/- 220.4      53.8   explicit[s=32]
   avg     32.3    312.6    151.6    158.8     39.1    398.0     84.0    101.8    115.8      1.2    1395.2     964.9+/- 228.6      32.7   explicit[s=64]

master (538ac87)
Total Timing Summary
Average time for map loading = 341.615 ms
Average time for input loading = 200098 ms
Average time for lst::Event creation = 0.00371252 ms
   Evt    Hits       MD       LS      T3       T5       pLS       pT5      pT3      TC       Reset    Event     Short             Rate
   avg     18.4    261.5    133.1    132.6     30.5    333.6     70.6     85.6     98.6      0.8    1165.3     813.3+/- 194.7    1166.5   explicit[s=1]
   avg     18.1    260.5    131.9    130.6     30.8    332.8     72.4     85.7     99.1      0.8    1162.8     811.9+/- 194.4     296.0   explicit[s=4]
   avg     20.5    288.7    143.6    144.8     34.4    366.7     79.3     94.7    108.6      1.0    1282.2     895.0+/- 214.1      93.9   explicit[s=16]
   avg     22.6    295.2    147.8    150.2     35.9    374.4     80.7     97.7    111.3      1.0    1316.8     919.8+/- 217.2      53.5   explicit[s=32]
   avg     32.4    312.6    152.0    159.1     39.3    405.7     83.9    102.2    119.3      1.1    1407.5     969.5+/- 230.3      32.8   explicit[s=64]
```