@slava77 2025-04-01
> The cpu timings on cgpu-1 also look good.

was this tested with `CUT_VALUE_DEBUG` enabled?