@slava77 2025-04-02
my review of the /standalone changes was rather cursory.
Perhaps a second look from others would help as well.

It would be good to understand what the impact is for the standard benchmark outputs.
Do I recall correctly that the timing benchmark is done without writing the ntuple?

How much slower is the case with an ntuple when default the validation plots are made (and what is the file size increase in this case)? Same for the case with `CUTVALUE_DEBUG` enabled.