@ariostas 2025-04-02
Thanks for all the comments, Slava!

> was this tested with CUT_VALUE_DEBUG enabled?

No, but I'll leave timings with it enabled below.

> a related question: is there a way to ask for it from the CI /run ...?

No, but if it becomes a common need I can add that. In that case we should also add the option to `lst_timing`.

> Perhaps a second look from others would help as well.

I'll also take a look. While rebasing I fixed some obvious things, but I didn't look at it very carefully.

> It would be good to understand what the impact is for the standard benchmark outputs.
> Do I recall correctly that the timing benchmark is done without writing the ntuple?

That's right. I'll also run the timing comparison with writing enabled.



I'll address some of the comments you left, but I'll refer the rest to @Hoobidoobidoo or @sgnoohc since they can better address them.

Heres a timing comparison on GPU with `CUTVALUE_DEBUG` enabled.
```
This PR (357dc9f)
Total Timing Summary
Average time for map loading = 548.513 ms
Average time for input loading = 7668.05 ms
Average time for lst::Event creation = 0.00314485 ms
   Evt    Hits       MD       LS      T3       T5       pLS       pT5      pT3      TC       Reset    Event     Short             Rate
   avg      1.3      0.5      0.4      1.8      0.6      0.5      0.9      0.5      1.1      0.0       7.6       5.8+/-  1.1       9.6   explicit_cutvalue[s=1]
   avg      1.3      0.7      0.6      2.2      0.8      0.7      1.2      0.7      1.5      0.0       9.6       7.6+/-  1.6       5.9   explicit_cutvalue[s=2]
   avg      2.3      1.1      1.1      3.1      1.6      1.1      2.2      1.1      2.5      0.0      16.0      12.7+/-  2.8       4.6   explicit_cutvalue[s=4]
   avg      3.7      1.8      1.7      4.2      2.6      1.6      3.1      1.5      3.8      0.0      23.8      18.6+/-  4.0       4.5   explicit_cutvalue[s=6]
   avg      5.5      2.7      2.4      5.2      3.6      2.0      4.4      2.0      4.9      0.0      32.7      25.3+/-  7.0       4.5   explicit_cutvalue[s=8]

master (538ac87)
Total Timing Summary
Average time for map loading = 547.471 ms
Average time for input loading = 7653.65 ms
Average time for lst::Event creation = 0.00329245 ms
   Evt    Hits       MD       LS      T3       T5       pLS       pT5      pT3      TC       Reset    Event     Short             Rate
   avg      1.3      0.5      0.4      1.8      0.6      0.5      0.9      0.5      1.1      0.0       7.6       5.8+/-  1.1       9.6   explicit_cutvalue[s=1]
   avg      1.3      0.7      0.6      2.1      0.8      0.7      1.1      0.7      1.5      0.0       9.5       7.5+/-  1.6       5.9   explicit_cutvalue[s=2]
   avg      2.2      1.2      1.1      3.1      1.6      1.1      2.2      1.1      2.5      0.0      16.1      12.8+/-  2.8       4.7   explicit_cutvalue[s=4]
   avg      3.5      1.9      1.7      4.1      2.5      1.6      3.2      1.4      3.7      0.0      23.7      18.6+/-  4.7       4.5   explicit_cutvalue[s=6]
   avg      5.5      2.8      2.3      5.1      3.6      2.1      3.9      2.0      4.9      0.0      32.1      24.5+/-  5.7       4.4   explicit_cutvalue[s=8]
```