@ariostas 2025-04-02
Here is a timing comparison with `CUTVALUE_DEBUG` and writing enabled.

```
This PR (357dc9f)
Total Timing Summary
Average time for map loading = 590.796 ms
Average time for input loading = 7655.72 ms
Average time for lst::Event creation = 0.00303133 ms
   Evt    Hits       MD       LS      T3       T5       pLS       pT5      pT3      TC       Reset    Event     Short             Rate
   avg      1.6      0.5      0.4      1.8      0.7      0.5      0.9      0.5      1.1      0.1       8.1       6.0+/-  1.2    2673.2   explicit_cutvalue[s=1]
   avg      1.4      0.5      0.4      1.8      0.7      0.5      1.0      0.6      1.1      0.1       8.0       6.1+/-  1.4    2857.4   explicit_cutvalue[s=2]
   avg      1.8      0.5      0.5      1.9      0.7      0.6      1.1      0.6      1.1      0.1       8.9       6.5+/-  3.2    2837.8   explicit_cutvalue[s=4]
   avg      2.5      0.6      0.6      2.0      0.8      0.6      1.3      0.8      1.3      0.1      10.5       7.4+/-  5.8    3140.6   explicit_cutvalue[s=6]
   avg      2.7      0.7      0.5      2.1      0.9      0.6      1.8      0.8      1.3      0.1      11.5       8.2+/-  9.6    2724.8   explicit_cutvalue[s=8]

master (538ac87)
Total Timing Summary
Average time for map loading = 568.772 ms
Average time for input loading = 8300.4 ms
Average time for lst::Event creation = 0.00299726 ms
   Evt    Hits       MD       LS      T3       T5       pLS       pT5      pT3      TC       Reset    Event     Short             Rate
   avg      3.1      0.6      0.6      2.0      0.8      0.6      1.5      1.1      1.2      0.1      11.7       8.0+/-  1.5    2859.7   explicit_cutvalue[s=1]
   avg      2.0      0.6      0.6      2.1      0.9      0.6      1.4      0.9      2.2      0.1      11.5       8.8+/-  3.5    2592.4   explicit_cutvalue[s=2]
   avg      4.5      0.8      0.8      2.3      1.1      0.8      2.4      2.4      3.8      0.2      19.1      13.8+/-  6.7    2875.7   explicit_cutvalue[s=4]
   avg      3.3      0.8      0.7      2.2      0.9      0.7      2.3      1.6      2.2      0.1      14.6      10.7+/-  7.3    2565.5   explicit_cutvalue[s=6]
   avg      4.0      0.9      0.7      2.4      1.1      0.8      2.5      1.6      2.2      0.1      16.0      11.3+/- 10.4    2604.3   explicit_cutvalue[s=8]
```