@ariostas 2025-07-25
The "standard" timing doesn't change significantly because it writes about the same data as before (although it increases enough to be a bottleneck for multiple streams). To enable the extra output you compile it with `-d` and it that case it does slow down significantly. @GNiendorf would you be able to check the notebooks since you're more familiar with them?

Here are timing tests on cgpu-1.

This is with `-w 1`.
```
master
Total Timing Summary
Average time for map loading = 594.484 ms
Average time for input loading = 16018.3 ms
Average time for lst::Event creation = 0.000715256 ms
   Evt    Hits       MD       LS      T3       T5       pLS       pT5      pT3      TC       Reset    Event     Short             Rate
   avg     14.0      0.5      0.4      0.7      0.8      0.7      0.8      0.4      1.1      0.0      19.5       4.9+/-  0.9     207.1   explicit[s=1]
   avg      1.3      0.5      0.4      0.7      0.8      0.7      0.9      0.5      1.1      0.0       7.0       5.0+/-  1.4     141.6   explicit[s=2]
   avg      1.6      0.5      0.5      0.8      0.9      0.7      1.1      0.5      1.1      0.0       7.8       5.4+/-  3.4     139.1   explicit[s=4]
   avg      2.3      0.6      0.5      0.9      1.0      0.7      1.4      0.6      1.2      0.0       9.1       6.1+/-  6.2     138.8   explicit[s=6]
   avg      2.7      0.7      0.5      0.9      1.0      0.7      1.4      0.8      1.3      0.0      10.0       6.6+/-  7.7     141.1   explicit[s=8]


This PR
Total Timing Summary
Average time for map loading = 592.162 ms
Average time for input loading = 7877.47 ms
Average time for lst::Event creation = 0.000669843 ms
   Evt    Hits       MD       LS      T3       T5       pLS       pT5      pT3      TC       Reset    Event     Short             Rate
   avg     14.8      0.5      0.4      0.7      0.8      0.6      0.9      0.4      1.1      0.0      20.3       4.9+/-  0.9     206.5   explicit[s=1]
   avg      1.2      0.5      0.4      0.8      0.8      0.7      0.9      0.5      1.1      0.0       6.9       5.0+/-  1.3     196.4   explicit[s=2]
   avg      1.5      0.5      0.5      0.8      0.9      0.7      1.1      0.5      1.1      0.0       7.6       5.4+/-  3.3     201.4   explicit[s=4]
   avg      2.0      0.6      0.5      0.9      1.0      0.7      1.3      0.6      1.2      0.0       8.7       6.0+/-  5.8     208.8   explicit[s=6]
   avg      2.6      0.6      0.5      1.0      1.1      0.8      1.8      0.6      1.3      0.0      10.4       7.0+/-  9.2     217.8   explicit[s=8]
```

This is compiled with `-d` and running it with `-w 1`.
```
master
Total Timing Summary
Average time for map loading = 593.076 ms
Average time for input loading = 7828.69 ms
Average time for lst::Event creation = 0.000669842 ms
   Evt    Hits       MD       LS      T3       T5       pLS       pT5      pT3      TC       Reset    Event     Short             Rate
   avg     15.2      0.5      0.4      0.8      0.8      0.7      1.0      0.5      1.1      0.1      21.1       5.2+/-  0.9    2769.5   explicit_cutvalue[s=1]
   avg      1.2      0.5      0.4      0.8      0.8      0.7      0.9      0.5      1.1      0.1       6.9       5.1+/-  1.2    2596.9   explicit_cutvalue[s=2]
   avg      1.6      0.5      0.5      0.8      0.9      0.7      1.1      0.5      1.1      0.1       7.7       5.4+/-  3.0    2608.1   explicit_cutvalue[s=4]
   avg      2.2      0.6      0.5      0.9      1.0      0.7      1.1      0.7      1.2      0.1       8.8       6.0+/-  5.2    2770.6   explicit_cutvalue[s=6]
   avg      2.6      0.6      0.5      0.9      1.1      0.7      1.5      0.7      1.3      0.1      10.1       6.7+/-  7.9    2895.8   explicit_cutvalue[s=8]

This PR
Total Timing Summary
Average time for map loading = 595.089 ms
Average time for input loading = 7812.24 ms
Average time for lst::Event creation = 0.000669843 ms
   Evt    Hits       MD       LS      T3       T5       pLS       pT5      pT3      TC       Reset    Event     Short             Rate
   avg     14.5      0.5      0.4      0.7      0.8      0.7      0.9      0.4      1.1      0.1      20.2       5.0+/-  0.9    4725.2   explicit_cutvalue[s=1]
   avg      1.3      0.5      0.4      0.8      0.8      0.7      0.9      0.5      1.1      0.1       7.1       5.1+/-  1.4    4952.4   explicit_cutvalue[s=2]
   avg      1.7      0.5      0.5      0.9      0.9      0.7      1.1      0.5      1.1      0.1       7.9       5.6+/-  3.1    4911.3   explicit_cutvalue[s=4]
   avg      2.1      0.6      0.5      0.9      1.0      0.7      1.3      0.6      1.2      0.1       9.0       6.2+/-  5.9    4970.4   explicit_cutvalue[s=6]
   avg      3.8      0.7      0.5      1.0      1.1      0.8      1.8      0.8      1.3      0.1      11.9       7.4+/-  9.1    4982.9   explicit_cutvalue[s=8]
```