@ariostas 2025-08-29
Okay, hopefully everything is good now.

I fixed the issues Gavin pointed out, moved some extra branches behind flags, and fixed the memory leak.

Here's the new timing for the standard outputs on cgpu-1. The ntuple is now only about 10% larger.
```
This PR
Total Timing Summary
Average time for map loading = 582.945 ms
Average time for input loading = 18068.4 ms
Average time for lst::Event creation = 0.000579016 ms
   Evt    Hits       MD       LS      T3       T5       pLS       pT5      pT3      TC       Reset    Event     Short             Rate
   avg     16.7      0.4      0.5      1.0      1.1      0.7      0.9      0.5      1.1      0.0      23.0       5.6+/-  1.6      23.0   explicit[s=1]
   avg      2.6      0.9      0.8      1.4      1.4      0.7      1.5      0.8      1.8      0.0      12.0       8.6+/-  2.1       6.0   explicit[s=2]
   avg      3.0      1.7      1.9      2.7      2.5      1.0      2.6      1.1      3.1      0.0      19.6      15.6+/-  2.8       5.0   explicit[s=4]
   avg      4.1      2.4      3.0      3.8      4.1      1.5      4.0      1.7      4.6      0.0      29.2      23.7+/-  5.0       5.0   explicit[s=6]
   avg      5.2      3.5      4.2      5.5      5.5      1.7      5.1      2.4      5.8      0.0      38.9      32.0+/-  6.8       5.0   explicit[s=8]

master
Total Timing Summary
Average time for map loading = 579.196 ms
Average time for input loading = 7751.67 ms
Average time for lst::Event creation = 0.00055631 ms
   Evt    Hits       MD       LS      T3       T5       pLS       pT5      pT3      TC       Reset    Event     Short             Rate
   avg     15.9      0.5      0.5      1.0      1.1      0.7      0.9      0.5      1.1      0.0      22.1       5.6+/-  1.6      22.1   explicit[s=1]
   avg      2.6      0.9      0.9      1.4      1.5      0.7      1.5      0.8      1.7      0.0      12.0       8.7+/-  2.0       6.1   explicit[s=2]
   avg      2.9      1.6      1.9      2.6      2.6      1.0      2.6      1.2      3.1      0.0      19.5      15.6+/-  3.1       5.0   explicit[s=4]
   avg      4.0      2.6      2.9      3.9      4.1      1.4      4.0      1.7      4.6      0.0      29.2      23.8+/-  5.1       5.0   explicit[s=6]
   avg      5.1      3.4      4.1      5.5      5.6      1.7      4.7      2.4      5.8      0.0      38.3      31.6+/-  4.4       4.9   explicit[s=8]
```

/run gpu-all